\name{get_returns}
\alias{get_returns}
\title{Get something something something described here}
\usage{
  get_returns(portfolio, start = "2012-01-01",
    end = "2012-12-01", name = "portfolio",
    benchmark = "\%5EGSPC", adjust = FALSE)
}
\description{
  Get something something something described here
}

